fix: 🐛 Add router to DHCP server options
This fixes an issue with Ludus deployments where users adding additional machines to GOAD will have those machines get DHCP from the DC, which does not include a default route, and thus the machines cannot be provisioned by Ludus. This change finds the default route for the DC, and adds that as the router option for the DHCP server.

Signed-off-by: kernel-sanders <1490292+kernel-sanders@users.noreply.github.com>